1.Added Switch Flag now is in IDE or standard exe envirmoents

+++ Modules/mdlCommandLine.bas
@@ -4,14 +4,14 @@ Option Explicit
 Private Declare Function CallWindowProc _
                 Lib "user32" _
                 Alias "CallWindowProcW" (ByVal lpPrevWndFunc As Long, _
-                                         ByVal hwnd As Long, _
+                                         ByVal hWnd As Long, _
                                          ByVal Msg As Long, _
                                          ByVal wParam As Long, _
                                          ByVal lParam As Long) As Long
 
 Private Declare Function SetWindowLong _
                 Lib "user32" _
-                Alias "SetWindowLongW" (ByVal hwnd As Long, _
+                Alias "SetWindowLongW" (ByVal hWnd As Long, _
                                         ByVal nIndex As Long, _
                                         ByVal dwNewLong As Long) As Long
 
@@ -45,7 +45,7 @@ Public Const PM_ACTIVE = &HFFE
 Public Const PM_PLAY_LAST = &HFFD
 
 Private Declare Sub DragAcceptFiles _
-                Lib "shell32.dll" (ByVal hwnd As Long, _
+                Lib "shell32.dll" (ByVal hWnd As Long, _
                                    ByVal fAccept As Long)
 
 Private Declare Sub DragFinish Lib "shell32.dll" (ByVal hDrop As Long)
@@ -67,7 +67,7 @@ Public Declare Function LocalFree Lib "kernel32" (ByVal hMem As IntPtr) As Long
 
 Private Declare Function PostMessage _
                 Lib "user32" _
-                Alias "PostMessageW" (ByVal hwnd As Long, _
+                Alias "PostMessageW" (ByVal hWnd As Long, _
                                       ByVal wMsg As Long, _
                                       ByVal wParam As Long, _
                                       ByVal lParam As Long) As Long
@@ -78,12 +78,12 @@ Private Declare Function TerminateProcess _
 
 Public Declare Function GetCurrentProcessId Lib "kernel32" () As Long
 
-Private Declare Function SetForegroundWindow Lib "user32" (ByVal hwnd As Long) As Long
+Private Declare Function SetForegroundWindow Lib "user32" (ByVal hWnd As Long) As Long
 
 Private Declare Function GetForegroundWindow Lib "user32" () As Long
 
 Private Declare Function GetWindowThreadProcessId _
-                Lib "user32" (ByVal hwnd As Long, _
+                Lib "user32" (ByVal hWnd As Long, _
                               lpdwProcessId As Long) As Long
 
 Private Declare Function AttachThreadInput _
@@ -91,7 +91,7 @@ Private Declare Function AttachThreadInput _
                               ByVal idAttachTo As Long, _
                               ByVal fAttach As Long) As Long
 
-Private Declare Function BringWindowToTop Lib "user32" (ByVal hwnd As Long) As Long
+Private Declare Function BringWindowToTop Lib "user32" (ByVal hWnd As Long) As Long
 
 Public Declare Function lstrlen _
                Lib "kernel32" _
@@ -102,7 +102,104 @@ Public Declare Sub CopyMemory _
                Alias "RtlMoveMemory" (hpvDest As Any, _
                                       hpvSource As Any, _
                                       ByVal cbCopy As Long)
-         
+
+Public Const MAX_PATH = 260
+
+Public Declare Function OpenClipboard Lib "user32" (ByVal hWnd As Long) As Long
+
+Public Declare Function CloseClipboard Lib "user32" () As Long
+
+Public Declare Function GetClipboardData Lib "user32" (ByVal wFormat As Long) As Long
+
+Private Declare Function GetModuleFileName _
+                Lib "kernel32" _
+                Alias "GetModuleFileNameA" (ByVal hModule As Long, _
+                                            ByVal lpFileName As String, _
+                                            ByVal nSize As Long) As Long
+
+Private Declare Function LoadLibrary _
+                Lib "kernel32" _
+                Alias "LoadLibraryA" (ByVal lpLibFileName As String) As Long
+
+Private Declare Function GetProcAddress _
+                Lib "kernel32" (ByVal hModule As Long, _
+                                ByVal lpProcName As String) As Long
+
+Private Declare Function CallAsmCode _
+                Lib "user32" _
+                Alias "CallWindowProcA" (lpPrevWndFunc As Long, _
+                                         ByVal hWnd As Long, _
+                                         ByVal Msg As Long, _
+                                         ByVal wParam As Long, _
+                                         lParam As Long) As Long
+
+Private Declare Function SetProcessDpiAwareness Lib "Shcore.dll" (ByVal value As Long) As Long
+
+Public Function SetProcessDpiAwareness_hard()
+    SetProcessDpiAwareness 2&
+End Function
+
+Public Function IsSupportDIPSet() As Boolean
+
+    Dim lLib As Long, hAddress As Long
+    
+    lLib = LoadLibrary("Shcore.dll")
+    IsSupportDIPSet = False
+    If (lLib <> 0) Then
+        hAddress = GetProcAddress(lLib, "SetProcessDpiAwareness")
+
+        If (hAddress <> 0) Then
+            IsSupportDIPSet = True
+            Exit Function
+        End If
+
+    End If
+
+End Function
+
+Public Function GetIDEmode() As Boolean
+
+    Dim strFileName As String
+
+    Dim lngCount    As Long
+
+    strFileName = String(255, 0)
+    lngCount = GetModuleFileName(App.hInstance, strFileName, 255)
+    strFileName = Left(strFileName, lngCount)
+
+    If NameGet(strFileName) = "VB6.EXE" Then
+
+        GetIDEmode = True
+    Else
+
+        GetIDEmode = False
+
+    End If
+
+End Function
+
+Public Sub ReadDrapQueryFile(ByVal hDrop As Long)
+
+    Dim i As Long, Count As Long, buf As String, sbuf As String
+
+    Count = DragQueryFile(hDrop, i, 0&, 0&)
+    
+    For i = 0 To Count - 1
+        buf = Space$(255)
+        sbuf = Space$(255)
+        DragQueryFile hDrop, i, StrPtr(buf), 254
+        GetShortPathName StrPtr(buf), StrPtr(sbuf), 255
+
+        If (Not (InStr(1, sbuf, Chr(0)) = 0)) Then
+            sbuf = Mid(sbuf, 1, InStr(1, sbuf, Chr(0)) - 1)
+
+        End If
+
+        mdlPlaylist.AddFileToPlaylist sbuf
+    Next
+
+End Sub
+
 Public Sub InitialCommandLine()
 
     Dim argc As Long, argv As IntPtr, i As Long
@@ -148,12 +245,12 @@ Public Sub Clearhook()
 
 End Sub
 
-Function HookCallback(ByVal hwnd As Long, _
+Function HookCallback(ByVal hWnd As Long, _
                       ByVal lMsg As Long, _
                       ByVal wParam As Long, _
                       ByVal lParam As Long) As Long
 
-    Select Case hwnd
+    Select Case hWnd
 
         Case hHookWindow
             MessageProc lMsg, wParam, lParam
@@ -163,7 +260,7 @@ Function HookCallback(ByVal hwnd As Long, _
     
     End Select
 
-    HookCallback = CallWindowProc(lPrevWndProc, hwnd, lMsg, wParam, lParam)
+    HookCallback = CallWindowProc(lPrevWndProc, hWnd, lMsg, wParam, lParam)
 
 End Function
 
@@ -246,8 +343,8 @@ Public Sub MessageProc(lMsg As Long, wParam As Long, lParam As Long)
             bPid = GetWindowThreadProcessId(GetForegroundWindow(), ByVal 0&)
             nPid = GetCurrentProcessId()
             AttachThreadInput bPid, nPid, True
-            SetForegroundWindow frmMain.hwnd
-            BringWindowToTop frmMain.hwnd
+            SetForegroundWindow frmMain.hWnd
+            BringWindowToTop frmMain.hWnd
             AttachThreadInput bPid, nPid, False
             frmMain.SetFocus
 
@@ -265,8 +362,59 @@ Public Sub ExitProgram()
     StopPlay
     CloseFile
     Clearhook
-    TerminateProcess GetCurrentProcessId, 5
-    Shell "taskkill /F /PID " & GetCurrentProcessId
+
+    If (Not IsIDE) Then TerminateProcess GetCurrentProcessId, 5
+    If (Not IsIDE) Then Shell "taskkill /F /PID " & GetCurrentProcessId
     End
 
 End Sub
+
+Private Function CallAnyFunc(ByVal pFn As Long, _
+                             ByVal pParam As Long, _
+                             ByVal Count As Long) As Long
+
+    Dim CallAnyFuncCode(34) As Long, lRet As Long
+
+    CallAnyFuncCode(0) = &H53EC8B55
+    CallAnyFuncCode(1) = &HE8&
+    CallAnyFuncCode(2) = &HEB815B00
+    CallAnyFuncCode(3) = &H1000112C
+    CallAnyFuncCode(4) = &H114A938D
+    CallAnyFuncCode(5) = &H64521000
+    CallAnyFuncCode(6) = &H35FF&
+    CallAnyFuncCode(7) = &H89640000
+    CallAnyFuncCode(8) = &H25&
+    CallAnyFuncCode(9) = &H8B1FEB00
+    CallAnyFuncCode(10) = &HE80C2444
+    CallAnyFuncCode(11) = &H0&
+    CallAnyFuncCode(12) = &H53E98159
+    CallAnyFuncCode(13) = &H8D100011
+    CallAnyFuncCode(14) = &H119791
+    CallAnyFuncCode(15) = &HB8908910
+    CallAnyFuncCode(16) = &H33000000
+    CallAnyFuncCode(17) = &H558BC3C0
+    CallAnyFuncCode(18) = &H104D8B0C
+    CallAnyFuncCode(19) = &HEB8A148D
+    CallAnyFuncCode(20) = &HFC528D06
+    CallAnyFuncCode(21) = &HB4932FF
+    CallAnyFuncCode(22) = &H8BF675C9
+    CallAnyFuncCode(23) = &HD0FF0845
+    CallAnyFuncCode(24) = &H58F64
+    CallAnyFuncCode(25) = &H83000000
+    CallAnyFuncCode(26) = &H4D8B04C4
+    CallAnyFuncCode(27) = &H5B018914
+    CallAnyFuncCode(28) = &H10C2C9
+    CallAnyFuncCode(29) = &H58F64
+    CallAnyFuncCode(30) = &H83000000
+    CallAnyFuncCode(31) = &HC03304C4
+    CallAnyFuncCode(32) = &H89144D8B
+    CallAnyFuncCode(33) = &HC2C95B21
+    CallAnyFuncCode(34) = &H90900010
+    CallAnyFunc = CallAsmCode(CallAnyFuncCode(0), pFn, pParam, Count, lRet)
+
+    If CallAnyFunc <> lRet Then
+        CallAnyFunc = 0 '??????????,???????????????
+        Debug.Assert False '??????????,?????????????
+    End If
+
+End Function
